requirements.txt
requests==2.31.0
PyGithub==2.1.1
python-dotenv==1.0.0
cryptography==41.0.7
flask-cors==4.0.0
werkzeug==3.0.1
